requirements.txt
streamlit>=1.30.0
anthropic>=0.20.0
pypdf>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=1.0.0
lxml>=4.9.0
requests>=2.31.0
boto3>=1.26.0
extra-streamlit-components>=0.1.60
numpy>=1.26.0
plotly>=5.18.0
matplotlib
pdf2image
pymupdf
psycopg2-binary>=2.9.9
